runner/pyproject.toml
[tool.poetry]
name = "runner"
version = "0.1.0"
description = "ILLIXR runner"
authors = ["Samuel Grayson <sam@samgrayson.me>"]
license = "NCSA"

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.3.1"
pyyaml-include = "^1.2"
jsonschema = "^3.2.0"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pylint = "^2.5.3"
mypy = "^0.782"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
